requirements.txt
altair==4.1.0
altair-data-server==0.4.1
altair-saver==0.5.0
altair-viewer==0.3.0
appdirs==1.4.3
attrs==19.3.0
black==19.10b0
click==7.1.1
entrypoints==0.3
importlib-metadata==1.6.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
mrjob==0.7.2
numpy==1.18.2
pandas==1.0.3
pathspec==0.8.0
portpicker==1.3.1
PyICU==2.4.3
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
regex==2020.4.4
selenium==3.141.0
six==1.14.0
toml==0.10.0
toolz==0.10.0
tornado==6.0.4
typed-ast==1.4.1
urllib3==1.25.9
vega-datasets==0.8.0
zipp==3.1.0
